# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ParityLab API
  version: 0.1.0
  description: Deterministic, sandbox-only Stripe integration verification.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/auth/register:
    post:
      operationId: register
      description: Atomically creates an owner, organization, project, default environments, and opaque session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password, workspace_name, project_name]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                password: { type: string, format: password, writeOnly: true, minLength: 12, maxLength: 256 }
                workspace_name: { type: string, minLength: 1, maxLength: 120 }
                project_name: { type: string, minLength: 1, maxLength: 120 }
      responses:
        '201':
          description: Authenticated session. The opaque token is set only as a hardened HttpOnly cookie.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionView' }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                password: { type: string, format: password, writeOnly: true, minLength: 1, maxLength: 256 }
      responses:
        '200':
          description: Authenticated session with an opaque HttpOnly cookie.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionView' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/auth/logout:
    post:
      operationId: logout
      security: [{ cookieAuth: [] }]
      responses:
        '204':
          description: Session revoked and cookie cleared. Repeated logout is safe.
          headers:
            Set-Cookie: { schema: { type: string } }
        '403': { $ref: '#/components/responses/Error' }
  /v1/session:
    get:
      operationId: getSession
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Current authenticated identity and project scope.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionView' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/settings/project:
    get:
      operationId: getProjectSettings
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Current project settings.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectSettings' }
        '401': { $ref: '#/components/responses/Error' }
    patch:
      operationId: updateProjectSettings
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                retention_days: { type: integer, minimum: 1, maximum: 3650 }
      responses:
        '200':
          description: Persisted project settings.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectSettings' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/environments:
    get:
      operationId: listEnvironments
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Project-scoped environments.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/List'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Environment' } }
        '401': { $ref: '#/components/responses/Error' }
  /v1/environments/{id}/select:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: selectEnvironment
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Newly selected project environment.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Environment' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/findings:
    get:
      operationId: listFindings
      security: [{ cookieAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [open, resolved, all], default: all }
      responses:
        '200':
          description: Project-scoped persisted findings.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/List'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Finding' } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/findings/{id}/resolve:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: resolveFinding
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Persisted resolved finding.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Finding' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/findings/{id}/reopen:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: reopenFinding
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Persisted reopened finding.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Finding' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/notifications:
    get:
      operationId: listNotifications
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Project-scoped notifications.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/List'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Notification' } }
        '401': { $ref: '#/components/responses/Error' }
  /v1/notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Count of unread notifications marked read.
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties:
                  updated: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/notifications/{id}/read:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: markNotificationRead
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Persisted read notification.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Notification' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/connections:
    get:
      operationId: listConnections
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Sanitized project-scoped Stripe connections. Secret material is never returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/List'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/StripeConnection' } }
        '401': { $ref: '#/components/responses/Error' }
...
  /v1/runs:
    get:
      operationId: listRuns
      responses:
        '200':
          description: Persisted runs ordered newest first
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/List'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Run' }
    post:
      operationId: createRun
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [scenario_id]
              properties:
                scenario_id: { type: string }
                fault: { $ref: '#/components/schemas/Fault' }
      responses:
        '201':
          description: Completed deterministic run; replay uses the same response and Idempotent-Replayed header.
          headers:
            Idempotent-Replayed:
              schema: { type: string, enum: ['true'] }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/runs/{id}:
    parameters:
      - $ref: '#/components/parameters/RunID'
    get:
      operationId: getRun
      responses:
        '200':
          description: Run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/runs/{id}/events:
    parameters:
      - $ref: '#/components/parameters/RunID'
    get:
      operationId: getRunEvents
      description: Returns JSON by default. Accept text/event-stream to replay run.event messages followed by run.complete.
      responses:
        '200':
          description: Event list or Server-Sent Events stream
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/List'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Event' }
            text/event-stream:
              schema: { type: string }
        '404': { $ref: '#/components/responses/Error' }
  /v1/runs/{id}/report:
    parameters:
      - $ref: '#/components/parameters/RunID'
    get:
      operationId: getRunReport
      responses:
        '200':
          description: Deterministic evidence report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/connections/stripe/validate:
    post:
      operationId: validateStripeConnection
      description: Validates a sandbox-only secret with Stripe and stores it encrypted. The secret is never returned.
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [secret_key]
              properties:
                secret_key:
                  type: string
                  writeOnly: true
                  pattern: '^(sk|rk)_test_'
                sandbox_name:
                  type: string
                  minLength: 1
                  maxLength: 120
      responses:
        '201':
          description: Sanitized connected Stripe Sandbox identity.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StripeConnection' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/stripe/payment-intents:
    post:
      operationId: createStripePaymentIntentRun
      description: Creates a real Stripe Sandbox PaymentIntent and persists its duplicate-submission verification evidence.
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [connection_id, amount_minor, currency]
              properties:
                connection_id: { type: string, format: uuid }
                amount_minor: { type: integer, minimum: 1, maximum: 99999999 }
                currency: { type: string, pattern: '^[a-z]{3}$' }
      responses:
        '201':
          description: Persisted verification run; replay returns the original run with Idempotent-Replayed set.
          headers:
            Idempotent-Replayed:
              schema: { type: string, enum: ['true'] }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /hooks/stripe/{endpoint_token}:
    post:
      operationId: ingestStripeWebhook
      description: Verifies Stripe-Signature against the untouched body and rejects livemode events.
      parameters:
        - name: endpoint_token
          in: path
          required: true
          schema: { type: string }
        - name: Stripe-Signature
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Quickly acknowledged, including whether this event identifier was already seen.
          content:
            application/json:
              schema:
                type: object
                required: [received, duplicate, event]
                properties:
                  received: { const: true }
                  duplicate: { type: boolean }
                  event:
                    type: object
                    required: [id, type]
                    properties:
                      id: { type: string }
                      type: { type: string }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: paritylab_session
      description: Opaque session token. Only its SHA-256 digest is persisted server-side.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 255 }
    RunID:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: '^run_[0-9]{6,}$' }
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Stripe-style API error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    Fault:
      type: string
      enum: [none, duplicate, reorder, timeout, tamper]
    SessionView:
      type: object
      additionalProperties: false
      required: [authenticated, user, organization, project, expires_at]
      properties:
        authenticated: { const: true }
        user:
          type: object
          additionalProperties: false
          required: [id, email]
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
        organization:
          type: object
          additionalProperties: false
          required: [id, name, role]
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            role: { type: string, enum: [owner, admin, member, viewer] }
        project:
          allOf:
            - $ref: '#/components/schemas/ProjectSettings'
        expires_at: { type: string, format: date-time }
    ProjectSettings:
      type: object
      additionalProperties: false
      required: [id, name, retention_days]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        retention_days: { type: integer, minimum: 1, maximum: 3650 }
    Environment:
      type: object
      additionalProperties: false
      required: [id, name, kind, is_default]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string, enum: [local, sandbox, staging] }
        is_default: { type: boolean }
    Notification:
      type: object
      additionalProperties: false
      required: [id, kind, payload, created_at]
      properties:
        id: { type: string, format: uuid }
        run_id: { type: string }
        kind: { type: string }
        payload: { type: object, additionalProperties: true }
        read_at:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        created_at: { type: string, format: date-time }
    StripeConnection:
      type: object
      additionalProperties: false
      required: [id, stripe_account_id, sandbox_name, status, created_at]
      properties:
        id: { type: string, format: uuid }
        stripe_account_id: { type: string, pattern: '^acct_' }
        sandbox_name: { type: string }
        status: { const: connected }
        created_at: { type: string, format: date-time }
    List:
      type: object
      required: [object, data, has_more]
      properties:
        object: { const: list }
        data: { type: array, items: {} }
        has_more: { const: false }
...
    Run:
      type: object
      required: [id, scenario_id, scenario_name, fault, status, score, started_at, completed_at, duration_ms, event_count, finding_count, recovered, environment, stripe_object_id, merchant_order_id]
      properties:
        id: { type: string }
        scenario_id: { type: string }
        scenario_name: { type: string }
        fault: { $ref: '#/components/schemas/Fault' }
        status: { type: string, enum: [running, passed, failed] }
        score: { type: integer, minimum: 0, maximum: 100 }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        duration_ms: { type: integer, minimum: 0 }
        event_count: { type: integer, minimum: 0 }
        finding_count: { type: integer, minimum: 0 }
        recovered: { type: boolean }
        environment: { const: sandbox }
        stripe_object_id: { type: string }
        merchant_order_id: { type: string }
    Event:
      type: object
      required: [id, run_id, sequence, at, source, target, type, title, detail, status, latency_ms, checkpoint, trace_id]
      properties:
        id: { type: string }
        run_id: { type: string }
        sequence: { type: integer, minimum: 1 }
        at: { type: string, format: date-time }
        source: { type: string }
        target: { type: string }
        type: { type: string }
        title: { type: string }
        detail: { type: string }
        status: { type: string, enum: [healthy, diverged, recovered, blocked] }
        latency_ms: { type: integer, minimum: 0 }
        checkpoint: { type: string }
        trace_id: { type: string }
        evidence: { type: object, additionalProperties: true }
        is_duplicate: { type: boolean }
    Finding:
      type: object
      required: [id, severity, title, summary, cause, remediation, checkpoint, resolved]
      properties:
        id: { type: string }
        severity: { type: string, enum: [info, warning, critical] }
        title: { type: string }
        summary: { type: string }
        cause: { type: string }
        remediation: { type: string }
        checkpoint: { type: string }
        resolved: { type: boolean }
    Report:
      type: object
      required: [run, summary, verdict, assertions, findings, state, generated_at]
      properties:
        run: { $ref: '#/components/schemas/Run' }
        summary: { type: string }
        verdict: { type: string }
        assertions: { type: array, items: { type: object, additionalProperties: true } }
        findings: { type: array, items: { $ref: '#/components/schemas/Finding' } }
        state:
          type: object
          required: [stripe, webhook, merchant, balanced]
          properties:
            stripe: { type: string }
            webhook: { type: string }
            merchant: { type: string }
            balanced: { type: boolean }
        generated_at: { type: string, format: date-time }
...